refactor. include license text, remove duplication to make sonar happy
