+++ libraries/src/AWS.Lambda.Powertools.JMESPath/Functions/EvaluateMinMax.cs
@@ -14,21 +14,13 @@
  */
 
 using System.Collections.Generic;
-using System.Diagnostics;
 
 namespace AWS.Lambda.Powertools.JMESPath.Functions;
 
-internal sealed class SortFunction : BaseFunction
+internal static class EvaluateMinMax
 {
-    internal SortFunction()
-        : base(1)
+    internal static bool TryEvaluate(IList<IValue> args, IBinaryOperator binaryOperator, out IValue element)
     {
-    }
-
-    public override bool TryEvaluate(DynamicResources resources, IList<IValue> args, out IValue element)
-    {
-        Debug.Assert(Arity.HasValue && args.Count == Arity!.Value);
-
         var arg0 = args[0];
         if (arg0.Type != JmesPathType.Array)
         {
@@ -36,43 +28,43 @@ internal sealed class SortFunction : BaseFunction
             return false;
         }
 
-        if (arg0.GetArrayLength() <= 1)
-        {
-            element = arg0;
-            return true;
-        }
-
-        var isNumber1 = arg0[0].Type == JmesPathType.Number;
-        var isString1 = arg0[0].Type == JmesPathType.String;
-        if (!isNumber1 && !isString1)
+        if (arg0.GetArrayLength() == 0)
         {
             element = JsonConstants.Null;
             return false;
         }
 
-        var comparer = ValueComparer.Instance;
-
-        var list = new List<IValue>();
-        foreach (var item in arg0.EnumerateArray())
+        var isNumber = arg0[0].Type == JmesPathType.Number;
+        var isString = arg0[0].Type == JmesPathType.String;
+        if (!isNumber && !isString)
         {
-            var isNumber2 = item.Type == JmesPathType.Number;
-            var isString2 = item.Type == JmesPathType.String;
-            if (!(isNumber2 == isNumber1 && isString2 == isString1))
+            element = JsonConstants.Null;
+            return false;
+        }
+
+        var index = 0;
+        for (var i = 1; i < arg0.GetArrayLength(); ++i)
+        {
+            if (!(arg0[i].Type == JmesPathType.Number == isNumber &&
+                  arg0[i].Type == JmesPathType.String == isString))
             {
                 element = JsonConstants.Null;
                 return false;
             }
 
-            list.Add(item);
+            if (!binaryOperator.TryEvaluate(arg0[i], arg0[index], out var value))
+            {
+                element = JsonConstants.Null;
+                return false;
+            }
+
+            if (Expression.IsTrue(value))
+            {
+                index = i;
+            }
         }
 
-        list.Sort(comparer);
-        element = new ArrayValue(list);
+        element = arg0[index];
         return true;
-    }
-
-    public override string ToString()
-    {
-        return "sort";
     }
 }
